+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# v0.1.1 / 2016-02-11
+
+* Changed return color to RGBA.
+* Added transparent background image.
+
 # v0.1.0 / 2015-12-10
 
 Create project